WIP: Translation multi-manipulator
