+++ applications-dev/plugins/SofaQtQuickGUI/data/qml/SofaInteractors/UserInteractor_EditMode.qml
@@ -32,17 +32,11 @@ UserInteractor_CameraMode {
         addMousePressedMapping(Qt.LeftButton, function(mouse, sofaViewer) {
             var particle = sofaViewer.pickParticle(Qt.point(mouse.x, mouse.y));
             if(particle) {
-//                if (particle.sofaComponent)
-//                    SofaApplication.selectedComponent = particle.sofaComponent
-                var manipulator = SofaApplication.selectedManipulator
-                if (manipulator)
+                if (particle.sofaComponent)
+                    SofaApplication.selectedComponent = particle.sofaComponent
+                else if (particle.manipulator)
                 {
-                    console.error("Got particle. index: " + particle.particleIndex)
-                    console.error("Current manipulator is: " + manipulator.name)
-
-                    manipulator.visible = true
-                    manipulator.sofaObject = particle.sofaComponent
-                    manipulator.particleIndex = particle.particleIndex
+                    SofaApplication.selectedManipulator = particle.manipulator
                     if (manipulator.mousePressed)
                         manipulator.mousePressed(mouse, sofaViewer)
                     if(manipulator.mouseMoved)
