+++ applications-dev/plugins/SofaQtQuickGUI/data/qml/SofaManipulators/Manipulator.qml
@@ -22,20 +22,4 @@ import Manipulator 1.0
 
 Manipulator {
     id: root
-
-    property string name
-//    onPositionChanged: {
-//        for(var i = 0; i < manipulators.length; ++i)
-//            manipulators[i].position = position;
-//    }
-
-//    onOrientationChanged: {
-//        for(var i = 0; i < manipulators.length; ++i)
-//            manipulators[i].orientation = orientation;
-//    }
-
-//    onScaleChanged: {
-//        for(var i = 0; i < manipulators.length; ++i)
-//            manipulators[i].scale = scale;
-//    }
 }
