+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Manipulators/_Manipulator2D_Translation.h
