+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Manipulators/_Manipulator3D_Rotation.h
