+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Manipulators/_Manipulator3D_Translation.cpp
